Update auctioneer.py
Please let me know if I handle things not as they're supposed to, this is the first time I've written code that someone else will see. ;) This is also the first time I had to work with multiple files, and therefore I only edited this one for now. I have some practical questions, but I'll ask on Monday.

The buyers/sellers need a code. I was thinking of (a) adding this to the namedtuples in the buyers/sellers file so they also get a code instantly instead of at the auctioneer process, or (b) let the auctioneer transform the list of namedtuples to a dictionary (keys = buyer/seller codes, values = tuples). Didn't do any of them yet, since I wanted to discuss this first.

The code doesn't yet include the complications mentioned by Nick, such as a new auction when a seller drops out of the auction (because of time and not knowing yet whether the current code suffices).

--> Sandholm, T. (2002). eMediator: A next generation electronic commerce server.
        Computational Intelligence, 18(4), 656-676.

Based on the Sandholm (2002) article, Middle Price 50:50 is best suited for determining
the prices in double auctions. This also happens to be a simple way of determining the
auction winners.
Assumptions:
- Demand is higher than supply (as discussed on Monday);
- Multiple units of 1 (homogeneous) item;
- All winners pay/get the same price.
I found schemes that were way more interesting (especially when keeping the relationship
with the other LCfP projects in mind), but it didn't fit the assumptions of an auction for
multiple units of a single item in double sealed-bid auctions.

To keep it simple (for now), I followed these steps:
(1) Calculate Qs
(2) Calculate median Ps
(3) Calculate Qb
(2) Deleted the lowest bids from the buyers, so Qs = Qb
(3) Calculate median Pb
(4) Calculate auction price (average of Ps and Pb)
(5) Distribute the items to the highest bidders

I tested the function with this data:
    Bid = namedtuple('Bid', ['action', 'quantity', 'item_price'])
    
    buyers = [Bid("buy", 200, 0.45),
              Bid("buy", 250, 0.60),
              Bid("buy", 400, 0.57),
              Bid("buy", 150, 0.64)]
    
    sellers = [Bid("sell", 100, 0.80),
               Bid("sell", 80, 0.90),
               Bid("sell", 120, 0.79),
               Bid("sell", 100, 0.88),
               Bid("sell", 90, 0.84),
               Bid("sell", 50, 0.99),
               Bid("sell", 60, 0.92)]
    
    Results:
        There are 600 items for sale.
        Total demand is 1000
        Median seller price: 0.84
        Median buyer price: 0.6
        Final auction price: 0.72
        Buyer 0 gets 150 units.
        Buyer 1 gets 250 units.
        Buyer 2 gets 200 units.
        Buyer 3 gets nothing.